Use memory fs and manually write out files
I needed a way to iterate over all files that Webpack outputs.  That
was necessary (and better than iterating over a real filesystem)
because the manifest file output by the DllPlugin doesn’t show up in
the stats.json output.

Also allows the webpack config options to be a function to lazily defer
options creation.

Need to revisit to ensure caching and file deletion works as expected.